pocs/reindex/requirements.txt
azure-storage-blob
azure-identity
langchain
openai
azure.search
azure-search-documents==11.4.0b8
tiktoken # OpenAIEmbeddings
unstructured
